requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.22.0
Babel==2.4.0
billiard==3.3.0.23
celery==3.1.25
cffi==1.10.0
cryptography==1.8.1
decorator==4.0.11
Django==1.8.17
django-celery==3.2.1
django-solo==1.1.2
django-suit==0.2.25
django-tinymce==2.6.0
flower==0.9.1
gunicorn==19.7.1
idna==2.5
ipython==5.3.0
ipython-genutils==0.2.0
kombu==3.0.37
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycparser==2.17
Pygments==2.2.0
pyOpenSSL==16.2.0
pyparsing==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
slackclient==1.0.5
tornado==4.2
traitlets==4.3.2
wcwidth==0.1.7
websocket-client==0.40.0
